split file into files
