--- csEmail/style.qss.ZN8500
@@ -1,11 +1,0 @@
-QToolButton{
-	min-width:40px;
-	min-height:40px;
-}
-QToolButton:hover{
-
-        border-image:url(":/images/icons/send.png");
-}
-QToolButton:pressed{
-	color:rgb(255,0,0);
-}
